resolve assumptions about alignment of inputs
ldm and stm require the address to be aligned to 4 bytes.
As keys, plaintexts, and ciphertexts are unsigned char*, those are not guaranteed
to be aligned in that way.
Hence, I replaced the ldm/stm with normal ldr/str which don't require such
alignment. It does increase the code-size a bit, but hopefully that does not
impact performance too badly.