Fix typo in masked key schedule functions
The third argument (i.e. the second 128-bit key) was not taken into consideration. As a result, the function did not output the expected result when two different keys were specified for the key schedule (since the fixsliced representation processes two blocks at a time, the round keys can be computed either for a single 128-bit key if the two blocks are encrypted with the same key, or for two 128-bit keys if the two blocks are encrypted with two different keys).